refactor: 영상 리랜더링 해결
