+++ src/app/upload/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useRef, useState } from "react";
+import { useMemo, useEffect, useRef, useState } from "react";
 import { useRouter } from "next/navigation";
 import styles from "./Upload.module.scss";
 import { ActionButton } from "@/components/atoms";
@@ -12,6 +12,34 @@ export default function UploadPage() {
   const handleToggle = () => {
     setIsPublic((prev) => !prev);
   };
+  const [videoFile, setVideoFile] = useState<File | null>(null);
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const handleClickArea = () => {
+    inputRef.current?.click();
+  };
+
+  const handleVideoChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    if (file) {
+      setVideoFile(file);
+    }
+  };
+
+  const videoSrc = useMemo(() => {
+    if (!videoFile) return null;
+    return URL.createObjectURL(videoFile);
+  }, [videoFile]);
+
+  // 객체 URL 메모리 해제 (중요!)
+  useEffect(() => {
+    return () => {
+      if (videoSrc) {
+        URL.revokeObjectURL(videoSrc);
+      }
+    };
+  }, [videoSrc]);
+  const fileName = videoFile?.name ?? "선택되지 않음";
   return (
     <>
       {/* 아래 div는 메뉴바 공간임 */}
@@ -27,7 +55,18 @@ export default function UploadPage() {
         <div className={styles.uploadContainer}>
           <div className={styles.videoContainer}>
             <h1 className={styles.titleText}>와글 썸네일</h1>
-            <VideoPreview />
+            <input
+              type="file"
+              accept="video/*"
+              ref={inputRef}
+              onChange={handleVideoChange}
+              style={{ display: "none" }}
+            />
+            <VideoPreview
+              src={videoSrc}
+              fileName={fileName}
+              onClick={handleClickArea}
+            />
           </div>
           <div className={styles.inputContainer}>
             <h3 className={styles.smallTitle}>제목</h3>
@@ -47,7 +86,7 @@ export default function UploadPage() {
               isPublic={isPublic}
               onClick={handleToggle}
             />
-            ;<ActionButton className={styles.actionButton}>완료</ActionButton>
+            <ActionButton className={styles.actionButton}>완료</ActionButton>
           </div>
         </div>
       </div>
